conteos/requirements.txt
Flask==3.0.0
#flask-admin==1.6.1
git+https://github.com/g-normand/flask-admin@master
Flask-SQLAlchemy==3.1.1
SQLAlchemy-serializer==1.4.1